refactor: キャラクター情報 APIs をモジュール化
